// react-frontend/src/api/message-service/api.ts
export interface ConversationCreate {
    /**
     * 
     * @type {string}
     * @memberof ConversationCreate
     */
    'type': string;
    /**
     * 
     * @type {string}
     * @memberof ConversationCreate
     */
    'title'?: string | null;
    /**
     * 
     * @type {Array<ConversationMemberIn>}
     * @memberof ConversationCreate
     */
    'members': Array<ConversationMemberIn>;
}
/**
 * 
 * @export
 * @interface ConversationMemberIn
 */
export interface ConversationMemberIn {
    /**
     * 
     * @type {string}
     * @memberof ConversationMemberIn
     */
    'user_id': string;
    /**
     * 
     * @type {string}
     * @memberof ConversationMemberIn
     */
    'role'?: string | null;
}